increase performance of validation in line items
This is done by (a) not doing a join from orders to store since the foreign key is already available and (b) not doing a join from the join table to the stores table since just checking the join table is enough. This also saves us another join from order to store again because we again have the foreign key available.